Worked on HighSchoolSlot script.
Added a new type of object called SelectedSport that has data type int and string for year  and name.

worked on selection list that keeps track of selected sports.
Made some comments on YearAdvance and Sport Delete methods.
    connections still need to be made with those to a specific button.

Slightly modified pushNext and pushPrevious on a similar line to make the code neater and less redudant.