Updated HighSchoolSlot Script
Added TaggedName Array, that i will use to be able to reference the tagged names of the Selection Buttons.
Added Display Selections Button.
Added DisplaySections to AdvanceYear and Delete Sport Method.

Need to figure out how to adjust a year when decrementing when not looking at that sport. (i.e. looking at football, but deleting baseballs year 2 and 3, i have to set that value back to 1 from 3.
Recheck if i actually need to be modifying the selection index here, or if i want to be modifying the value at that index.