Updated HighschoolSlot Script
finished everything i had planned for the DeleteSport() method. I figured out how to deal with deleted not "current" sports.
I had to refer to their name in selections before i deleted it and then looked for that in the sports array. I use the index of that sport, which mirrors the same index of Sport Years and can decrement it then.